Day 1 learning
The concept of state is interesting. I would redo many things knowing this now. How object orientation really works is still new, I would put things in different places. What exists in an object and how it is passed and access to various methods is also a bit fuzzy at this point.